many, many cleanups
* agen5/ag-text.def: correct which file is to be removed
* agen5/agDep.c: prefer memcpy to strcpy
  (tidy_dep_file): prefer fchmod to chmod
* agen5/expOutput.c (make_writable): obsolete
* agen5/expString.c (stringify_for_sh): remove unused parameter
* agen5/funcFor.c (call_gen_block): new function to ensure longjmp works
* autoopts/test/defs.in (CFLAGS): -Werror may be suffixed.  Remove all.
  (clean_help): no need for "egrep -v" when sed can do the job
* autoopts/load.c (add_env_val): no need to fix -Werror here
* autoopts/tpl/opt*: remove texi-isms from help text.
* autoopts/usage.c (prt_extd_usage): arg types param is not used.
  (prt_one_vendor): thus, here, too.
  (prt_vendor_opts): and here as well.
* doc/mk-agen-texi-in: -Werror removal needed here, too.
* everywhere: warning cleanups & de-camelization
* agen5/Makefile.am: Remove lists of generated things and derive them
  from the script that has to know about them:  bootstrap.dir
* agen5/ag-text.def: make the touchback silent and fix the format numbers
* agen5/agDep.c (wrap_up_depends): use the fixed format correctly
* agen5/bootstrap.dir: edit Makefile.am, too
* agen5/fsm-macro.tlib: use the built-in temp directory
* agen5/fsm-trans.tlib: likewise
* config/bootstrap.shlib (make_opts) cleanup for dependency generation
  (make_proto): likewise
* doc/auto_gen-tpl.in: remove use of ADDON_MENU
* doc/mk-agen-texi-in: stop providing ADDON_MENU

+++ agen5/defLex.c
@@ -2,7 +2,7 @@
 /**
  * @file defLex.c
  *
- *  Time-stamp:        "2012-03-10 10:12:20 bkorb"
+ *  Time-stamp:        "2012-03-31 13:00:08 bkorb"
  *
  *  This module scans the template variable declarations and passes
  *  tokens back to the parser.
@@ -412,7 +412,7 @@ loadScheme(void)
     pzEnd = (char*)scm2display(res); /* ignore const-ness */
     cctx->scx_line = next_ln;
 
-    if (strlen(pzEnd) >= schemeLen) {
+    if ((int)strlen(pzEnd) >= schemeLen) {
         AGDUPSTR(pzEnd, pzEnd, "SCM Result");
 
         token_str = pzEnd;
@@ -559,7 +559,7 @@ gather_name(char * pzScan, te_dp_event * pRetVal)
                 *pRetVal = aKeywordTkn[ kw_ix ];
                 break;
             }
-        } while (++kw_ix < KEYWORD_CT);
+        } while (++kw_ix < (int)KEYWORD_CT);
 
         *pzScan = sv_ch;         /* restore the following character  */
     }
